+++ packages/create-dojo/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://github.com/dojoengine/dojo.js/compare/v1.0.3-alpha.2...v1.0.3) (2024-12-14)
+
+
+### Bug Fixes
+
+* create dojo contract manifest path ([a6fb293](https://github.com/dojoengine/dojo.js/commit/a6fb29322468a71df4621877f74ec280ce004be8))
+
+
+
+
+
 ## [1.0.2](https://github.com/dojoengine/dojo.js/compare/v1.0.1...v1.0.2) (2024-11-27)
 
 **Note:** Version bump only for package @dojoengine/create-dojo