+++ packages/sdk/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://github.com/dojoengine/sdk/compare/v1.0.3-alpha.2...v1.0.3) (2024-12-14)
+
+
+### Features
+
+* add order_by and entity_models to sdk ([0c9cf79](https://github.com/dojoengine/sdk/commit/0c9cf7913bc5b50bd907f56f3c60e169ef43ecce))
+* upgrade to torii v1.0.7 ([5966fcc](https://github.com/dojoengine/sdk/commit/5966fcc072b02ec49bba4770031bc4fd760ee14a))
+
+
+
+
+
 ## [1.0.2](https://github.com/dojoengine/sdk/compare/v1.0.1...v1.0.2) (2024-11-27)